Improved handling of new client connections.
New clients now receive the state of all other connected clients, instead of having to wait for each of the connected clients to click something and send a message.

The new client also receives a message for what its default state should be.

Previously connected clients also receive a message about the new client's state instead of having to wait for the new client to click something.

This also fixes a bug from the previous commit where, if a new client connects without receiving a message from another client, and that other client disconnects, the new client's page crashes.

While functionality looks great, this code is in heavy need of refactoring.